Documents/Cursor/requirements.txt
anthropic==0.18.1
pyautogui==0.9.54
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
python-dotenv==1.0.1 